Discount with TRGO points

+++ lib/Controllers/BookingBackend/hotel_booking.dart
@@ -69,7 +69,7 @@ class HotelBooking {
       int numberOfAdult,
       int numberOfChildren,
       String room,
-      int price,
+      double price,
       int age,
       String bookingId) async {
     try {
